GUI Improvements
I think I finished the GUI for the most part. If you guys suggest
adds/edits to it , let me know. I believe the next step, is to work on
re-implementing the accumulate and calculate expression parts. I have it
where if it is on the proper "mode" if someone press enter with values
put in, it should try to call the function associated with that mode.