May Break Code
So Bowman finally got back to me on making packages, it's a no-no. I
thought it was needed in order to use other classes but if it's in the
same directory it should be ok. So I am attempting to take out the
packages and have it still work. Since it's removing folders and
whatnot, it may cause problems